:sparkles: add handler for FkOrURLField
